+++ matchms/filtering/__init__.py
@@ -55,7 +55,7 @@ from .harmonize_undefined_inchi import harmonize_undefined_inchi
 from .harmonize_undefined_inchikey import harmonize_undefined_inchikey
 from .harmonize_undefined_smiles import harmonize_undefined_smiles
 from .interpret_pepmass import interpret_pepmass
-from .load_adducts import load_adducts_dict, load_known_adduct_conversions
+from .repair_adduct.clean_adduct import load_adducts_dict, load_known_adduct_conversions
 from .make_charge_int import make_charge_int
 from .make_charge_scalar import make_charge_scalar
 from .make_ionmode_lowercase import make_ionmode_lowercase
@@ -96,8 +96,6 @@ __all__ = [
     "harmonize_undefined_inchikey",
     "harmonize_undefined_smiles",
     "interpret_pepmass",
-    "load_adducts_dict",
-    "load_known_adduct_conversions",
     "make_charge_int",
     "make_charge_scalar",
     "make_ionmode_lowercase",
