+++ matchms/filtering/filter_utils/derive_precursor_mz_and_parent_mass.py
@@ -1,8 +1,7 @@
 import logging
 
-from matchms.filtering.load_adducts import load_adducts_dict
 from matchms.constants import PROTON_MASS
-from matchms.filtering.repair_adduct.clean_adduct import clean_adduct
+from matchms.filtering.repair_adduct.clean_adduct import clean_adduct, load_adducts_dict
 
 logger = logging.getLogger("matchms")
