+++ tests/test_metadata_utils.py
@@ -6,12 +6,11 @@ from unittest import mock
 import numpy as np
 import pytest
 import matchms.metadata_utils
-from matchms.metadata_utils import (clean_adduct,
-                                    derive_fingerprint_from_inchi,
+from matchms.metadata_utils import (derive_fingerprint_from_inchi,
                                     derive_fingerprint_from_smiles,
                                     is_valid_inchi, is_valid_inchikey,
-                                    is_valid_smiles, looks_like_adduct,
-                                    mol_converter)
+                                    is_valid_smiles, mol_converter)
+from matchms.filtering.repair_adduct.clean_adduct import looks_like_adduct, clean_adduct
 
 
 @pytest.fixture()
